+++ app/src/main/kotlin/me/amrbashir/hijriwidget/HijriDate.kt
@@ -2,89 +2,61 @@ package me.amrbashir.hijriwidget
 
 import android.icu.util.Calendar
 import android.icu.util.ULocale
-import android.util.Log
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 
+object HijriDate {
+    val today: MutableState<String> = mutableStateOf("")
 
-val AR_MONTHS = arrayOf(
-    "مُحَرَّم",
-    "صَفَر",
-    "رَبِيع الأوَّل",
-    "رَبِيع الثَّاني",
-    "جُمَادى الأولى",
-    "جُمَادى الآخرة",
-    "رَجَب",
-    "شَعْبَان",
-    "رَمَضَان",
-    "شَوَّال",
-    "ذُو الْقَعْدَة",
-    "ذُو الْحِجَّة"
-)
-val EN_MONTHS = arrayOf(
-    "Muharram",
-    "Safar",
-    "Rabi’ al-Awwal",
-    "Rabi’ al-Thani",
-    "Jumada al-Awwal",
-    "Jumada al-Thani",
-    "Rajab",
-    "Sha’ban",
-    "Ramadan",
-    "Shawwal",
-    "Dhu al-Qi’dah",
-    "Dhu al-Hijjah",
-)
-
-
-class HijriDate(val day: Int, val month: Int, val year: Int) {
-
-    fun display(): String {
-        val lang = Preferences.language.value
-
-        val day = day.toString().convertNumbersToLang(Preferences.language.value)
-        val monthName = when (lang) {
-            SupportedLanguage.English -> EN_MONTHS[month]
-            else -> AR_MONTHS[month]
-        }
-        val year = year.toString().convertNumbersToLang(Preferences.language.value)
-
-        return "$day $monthName $year"
+    fun load() {
+        this.today.value = todayStr()
     }
 
-    companion object {
-        val today: MutableState<HijriDate> = mutableStateOf(HijriDate(0, 0, 0))
+    fun today(): Calendar {
+        val locale = ULocale("@calendar=${Preferences.calendarCalculationMethod.value}")
+        val calendar = Calendar.getInstance(locale)
 
-        fun load() {
-            this.today.value = today()
+        val dayStart = Preferences.dayStart.value
+
+        calendar.add(Calendar.DAY_OF_MONTH, Preferences.dayOffset.value)
+
+        val currentMinutes = calendar[Calendar.HOUR_OF_DAY] * 60 + calendar[Calendar.MINUTE]
+        val dayStartMinutes = dayStart.hour * 60 + dayStart.minute
+
+        if (currentMinutes < dayStartMinutes) {
+            calendar.add(Calendar.DAY_OF_MONTH, -1)
         }
 
-        fun today(): HijriDate {
-            val dayStart = Preferences.dayStart.value
+        return calendar
+    }
 
-            val calendar =
-                Calendar.getInstance(ULocale("@calendar=${Preferences.calendarCalculationMethod.value}"))
+    fun todayStr(): String {
+        val today = today()
+        return if (Preferences.isCustomFormat.value) Preferences.customFormat.value.formatDate(today)
+        else Preferences.format.value.formatDate(today)
+    }
 
-            Log.d(
-                "HijriDate",
-                "Day before offset: ${calendar[Calendar.DAY_OF_MONTH]}/${calendar[Calendar.MONTH]}"
-            )
-            calendar.add(Calendar.DAY_OF_MONTH, Preferences.dayOffset.value)
-            Log.d(
-                "HijriDate",
-                "Day after offset: ${calendar[Calendar.DAY_OF_MONTH]}/${calendar[Calendar.MONTH]}"
-            )
-
-            val dayOfMonth = calendar[Calendar.DAY_OF_MONTH]
-            val day = if (
-                calendar[Calendar.HOUR_OF_DAY] >= dayStart.hour &&
-                calendar[Calendar.MINUTE] >= dayStart.minute
-            ) dayOfMonth else dayOfMonth - 1
-
-            val month = calendar[Calendar.MONTH]
-            val year = calendar[Calendar.YEAR]
-
-            return HijriDate(day, month, year)
-        }
+    fun todayNumber(): Int {
+        return today()[Calendar.DAY_OF_MONTH]
     }
 }
+
+
+enum class HijriDateCalculationMethod(val id: String, val label: String, val description: String) {
+    ISLAMIC(
+        "islamic",
+        "Isalmic",
+        "Uses pure astronomical calculations to determine when the new moon occurs"
+    ),
+    ISLAMIC_UMALQURA(
+        "islamic-umalqura",
+        "Islamic Umm al-Qura",
+        "Based on the calculations used by the Umm al-Qura University in Mecca"
+    ),
+    ISLAMIC_CIVIL(
+        "islamic-civil",
+        "Islamic Civil",
+        "Uses a fixed algorithmic approach with a 30-year cycle"
+    ),
+}
+