+++ app/src/main/kotlin/me/amrbashir/hijriwidget/Preferences.kt
@@ -1,8 +1,6 @@
 package me.amrbashir.hijriwidget
 
 import android.content.Context
-import android.content.res.Configuration
-import android.content.res.Configuration.UI_MODE_NIGHT_YES
 import android.icu.util.Calendar
 import android.os.Build
 import androidx.compose.material3.dynamicDarkColorScheme
@@ -17,7 +15,6 @@ import java.time.LocalTime
 import java.time.format.DateTimeFormatter
 
 private const val PREF = "HijriWidgetPref"
-private const val LANG_KEY = "LANG"
 private const val THEME_KEY = "THEME"
 private const val CUSTOM_COLOR_KEY = "CUSTOM_COLOR"
 private const val BG_THEME_KEY = "BG_THEME"
@@ -28,47 +25,54 @@ private const val DAY_START_HOUR_KEY = "DAY_START_HOUR"
 private const val DAY_START_MINUTE_KEY = "DAY_START_MINUTE"
 private const val DAY_OFFSET_KEY = "DAY_OFFSET"
 private const val CALENDAR_CALCULATION_METHOD_KEY = "CALENDAR_CALCULATION_METHOD"
+private const val FORMAT_KEY = "FORMAT"
+private const val IS_CUSTOM_FORMAT_KEY = "IS_CUSTOM_FORMAT"
+private const val CUSTOM_FORMAT_KEY = "CUSTOM_FORMAT"
 
 object Preferences {
-    val language: MutableState<SupportedLanguage> = mutableStateOf(Defaults.language)
     val theme: MutableState<SupportedTheme> = mutableStateOf(Defaults.theme)
     val color: MutableState<Int> = mutableIntStateOf(Defaults.color)
     val customColor: MutableState<Int> = mutableIntStateOf(Defaults.color)
     val bgTheme: MutableState<SupportedTheme> = mutableStateOf(Defaults.bgTheme)
     val bgColor: MutableState<Int> = mutableIntStateOf(Defaults.bgColor)
     val bgCustomColor: MutableState<Int> = mutableIntStateOf(Defaults.bgColor)
-    val customTextSize: MutableState<Float> = mutableFloatStateOf(Defaults.customTextSize)
+    val textSize: MutableState<Float> = mutableFloatStateOf(Defaults.textSize)
     val shadow: MutableState<Boolean> = mutableStateOf(Defaults.shadow)
     val dayStart: MutableState<DayStart> = mutableStateOf(Defaults.dayStart)
     val dayOffset: MutableState<Int> = mutableIntStateOf(Defaults.dayOffset)
     val calendarCalculationMethod: MutableState<String> =
         mutableStateOf(Defaults.calendarCalculationMethod)
+    val format: MutableState<String> = mutableStateOf(Defaults.format)
+    val isCustomFormat: MutableState<Boolean> = mutableStateOf(Defaults.isCustomFormat)
+    val customFormat: MutableState<String> = mutableStateOf(Defaults.customFormat)
 
     @Suppress("ConstPropertyName")
     object Defaults {
-        val language = SupportedLanguage.Arabic
         val theme =
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) SupportedTheme.Dynamic else SupportedTheme.System
         val color = Color.White.toArgb()
-        val bgTheme =
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) SupportedTheme.Dynamic else SupportedTheme.Transparent
+        val bgTheme = SupportedTheme.Transparent
         val bgColor = Color.Transparent.toArgb()
-        const val customTextSize = 22F
+        const val textSize = 22F
         const val shadow = true
         val dayStart = DayStart(0, 0)
         const val dayOffset = 0
-        val calendarCalculationMethod = CalendarCalculationMethod.ISLAMIC_UMALQURA.toString()
+        val calendarCalculationMethod = HijriDateCalculationMethod.ISLAMIC_UMALQURA.id
+        val format = FORMAT_PRESETES[0]
+        const val isCustomFormat = false
+        const val customFormat = ""
     }
 
     fun restoreDefaults(context: Context) {
-        this.language.value = Defaults.language
         this.theme.value = Defaults.theme
         this.bgTheme.value = Defaults.bgTheme
-        this.customTextSize.value = Defaults.customTextSize
+        this.textSize.value = Defaults.textSize
         this.shadow.value = Defaults.shadow
         this.dayStart.value = Defaults.dayStart
         this.dayOffset.value = Defaults.dayOffset
         this.calendarCalculationMethod.value = Defaults.calendarCalculationMethod
+        this.format.value = Defaults.format
+        this.isCustomFormat.value = Defaults.isCustomFormat
         this.updateColor(context)
     }
 
@@ -76,21 +80,20 @@ object Preferences {
     fun load(context: Context) {
         val sharedPreferences = context.getSharedPreferences(PREF, 0)
 
-        val lang = sharedPreferences.getString(LANG_KEY, "Arabic") ?: "Arabic"
-        this.language.value = SupportedLanguage.valueOf(lang)
-
-        val theme = sharedPreferences.getString(THEME_KEY, "Dynamic") ?: "Dynamic"
+        val theme = sharedPreferences.getString(THEME_KEY, Defaults.theme.toString())
+            ?: Defaults.theme.toString()
         this.theme.value = SupportedTheme.valueOf(theme)
         this.customColor.value = sharedPreferences.getInt(CUSTOM_COLOR_KEY, Color.White.toArgb())
 
-        val bgTheme = sharedPreferences.getString(BG_THEME_KEY, "Dynamic") ?: "Dynamic"
+        val bgTheme = sharedPreferences.getString(BG_THEME_KEY, Defaults.bgTheme.toString())
+            ?: Defaults.bgTheme.toString()
         this.bgTheme.value = SupportedTheme.valueOf(bgTheme)
         this.bgCustomColor.value =
             sharedPreferences.getInt(BG_CUSTOM_COLOR_KEY, Color.Transparent.toArgb())
 
         this.shadow.value = sharedPreferences.getBoolean(SHADOW_KEY, true)
 
-        this.customTextSize.value = sharedPreferences.getFloat(CUSTOM_TEXT_SIZE_KEY, 22F)
+        this.textSize.value = sharedPreferences.getFloat(CUSTOM_TEXT_SIZE_KEY, 22F)
 
         this.dayStart.value = DayStart(
             sharedPreferences.getInt(DAY_START_HOUR_KEY, 0),
@@ -100,8 +103,17 @@ object Preferences {
         this.dayOffset.value = sharedPreferences.getInt(DAY_OFFSET_KEY, 0)
 
         this.calendarCalculationMethod.value =
-            sharedPreferences.getString(CALENDAR_CALCULATION_METHOD_KEY, "islamic-umalqura")
-                ?: "islamic-umalqura"
+            sharedPreferences.getString(
+                CALENDAR_CALCULATION_METHOD_KEY,
+                Defaults.calendarCalculationMethod
+            )
+                ?: Defaults.calendarCalculationMethod
+
+        this.format.value =
+            sharedPreferences.getString(FORMAT_KEY, Defaults.format) ?: Defaults.format
+        this.isCustomFormat.value =
+            sharedPreferences.getBoolean(IS_CUSTOM_FORMAT_KEY, Defaults.isCustomFormat)
+        this.customFormat.value = sharedPreferences.getString(CUSTOM_FORMAT_KEY, "") ?: ""
 
         this.updateColor(context)
         this.updateBgColor(context)
@@ -110,13 +122,12 @@ object Preferences {
     fun save(context: Context) {
         val sharedPreferences = context.getSharedPreferences(PREF, 0)
         sharedPreferences.edit()?.run {
-            putString(LANG_KEY, this@Preferences.language.value.toString())
             putString(THEME_KEY, this@Preferences.theme.value.toString())
             putInt(CUSTOM_COLOR_KEY, this@Preferences.customColor.value)
             putString(BG_THEME_KEY, this@Preferences.bgTheme.value.toString())
             putInt(BG_CUSTOM_COLOR_KEY, this@Preferences.bgCustomColor.value)
             putBoolean(SHADOW_KEY, this@Preferences.shadow.value)
-            putFloat(CUSTOM_TEXT_SIZE_KEY, this@Preferences.customTextSize.value)
+            putFloat(CUSTOM_TEXT_SIZE_KEY, this@Preferences.textSize.value)
             putInt(DAY_START_HOUR_KEY, this@Preferences.dayStart.value.hour)
             putInt(DAY_START_MINUTE_KEY, this@Preferences.dayStart.value.minute)
             putInt(DAY_OFFSET_KEY, this@Preferences.dayOffset.value)
@@ -124,6 +135,9 @@ object Preferences {
                 CALENDAR_CALCULATION_METHOD_KEY,
                 this@Preferences.calendarCalculationMethod.value
             )
+            putString(FORMAT_KEY, this@Preferences.format.value)
+            putBoolean(IS_CUSTOM_FORMAT_KEY, this@Preferences.isCustomFormat.value)
+            putString(CUSTOM_FORMAT_KEY, this@Preferences.customFormat.value)
             commit()
         }
     }
@@ -194,7 +208,3 @@ data class DayStart(val hour: Int, val minute: Int) {
     }
 }
 
-
-fun Context.isDark(): Boolean {
-    return resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK == UI_MODE_NIGHT_YES
-}