+++ app/src/main/kotlin/me/amrbashir/hijriwidget/preferences/LauncherIcon.kt
@@ -7,7 +7,7 @@ import me.amrbashir.hijriwidget.BuildConfig
 import me.amrbashir.hijriwidget.HijriDate
 
 fun changeLauncherIcon(context: Context) {
-    val today = HijriDate.today()
+    val today = HijriDate.todayNumber()
 
     val packageManager = context.packageManager
 
@@ -26,7 +26,7 @@ fun changeLauncherIcon(context: Context) {
             ComponentName(context, "me.amrbashir.hijriwidget.preferences.Calendar_$day")
         packageManager.setComponentEnabledSetting(
             dayActivity,
-            if (today.day == day) PackageManager.COMPONENT_ENABLED_STATE_ENABLED else PackageManager.COMPONENT_ENABLED_STATE_DISABLED,
+            if (today == day) PackageManager.COMPONENT_ENABLED_STATE_ENABLED else PackageManager.COMPONENT_ENABLED_STATE_DISABLED,
             PackageManager.DONT_KILL_APP
         )
     }